+++ src/equations/inviscid_burgers_1d.jl
@@ -80,12 +80,6 @@ end
   λ_max = max(abs(u_L), abs(u_R))
 end
 
-function flux_lax_friedrichs(u_ll, u_rr, orientation, equation::InviscidBurgersEquation1D)
-  u_L = u_ll[1]
-  u_R = u_rr[1]
-
-  return SVector(0.5 * ( 0.5 * (u_L^2 + u_R^2) - max(abs(u_L), abs(u_R)) * (u_R - u_L) ))
-end
 
 @inline function max_abs_speeds(u, equation::InviscidBurgersEquation1D)
   return (abs(u[1]),)