+++ src/equations/linear_scalar_advection_3d.jl
@@ -139,12 +139,6 @@ end
   λ_max = abs(equation.advectionvelocity[orientation])
 end
 
-function flux_lax_friedrichs(u_ll, u_rr, orientation, equation::LinearScalarAdvectionEquation3D)
-  a = equation.advectionvelocity[orientation]
-  return 0.5 * ( a * (u_ll + u_rr) - abs(a) * (u_rr - u_ll) )
-end
-
-
 
 @inline have_constant_speed(::LinearScalarAdvectionEquation3D) = Val(true)